Make QuadBuffer.expandCapacity() take a size_t instead of ulong
This is necessary for 32-bit builds to compile

+++ source/magra/renderer.d
@@ -117,7 +117,7 @@ class QuadBuffer
         program = new ShaderProgram;
     }
 
-    void expandCapacity(ulong extra)
+    void expandCapacity(size_t extra)
     {        
         auto oldEboLength = eboData.length;
